requirements.txt
alembic==1.7.3
blinker==1.4
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
shell==1.0.1
SQLAlchemy==1.4.23
visitor==0.1.3
Werkzeug==0.16.0
certifi==2021.5.30
charset-normalizer==2.0.4
pathlib==1.0.1
googlefonts-installer==0.3.1
urllib3==1.26.6
